+++ src/swiper/demos/base.vue
@@ -1,25 +1,27 @@
 <template>
-  <div class="tdesign-mobile-demo">
-    <div style="padding: 0 16px">
-      <t-swiper>
-        <t-swiper-item class="swiper-item--demo">
-          <img src="https://tdesign.gtimg.com/site/swiper/01.png" class="img" />
-        </t-swiper-item>
-        <t-swiper-item class="swiper-item--demo">
-          <img src="https://tdesign.gtimg.com/site/swiper/02.png" class="img" />
-        </t-swiper-item>
-        <t-swiper-item class="swiper-item--demo">
-          <img src="https://tdesign.gtimg.com/site/swiper/03.png" class="img" />
-        </t-swiper-item>
-      </t-swiper>
-    </div>
-  </div>
+  <t-swiper
+    :interval="3000"
+    :autoplay="true"
+    :navigation="{ type: 'dots', showSlideBtn: false }"
+    @change="handleChange"
+  >
+    <t-swiper-item class="swiper-item--demo">
+      <img src="https://tdesign.gtimg.com/site/swiper/01.png" class="img" />
+    </t-swiper-item>
+    <t-swiper-item class="swiper-item--demo">
+      <img src="https://tdesign.gtimg.com/site/swiper/02.png" class="img" />
+    </t-swiper-item>
+    <t-swiper-item class="swiper-item--demo">
+      <img src="https://tdesign.gtimg.com/site/swiper/03.png" class="img" />
+    </t-swiper-item>
+  </t-swiper>
 </template>
-<script lang="ts" setup></script>
+<script lang="ts" setup>
+const handleChange = (index: number) => {
+  console.log('基础示例,页数变化到》》》', index);
+};
+</script>
 <style lang="less">
-.t-swiper {
-  margin-bottom: 24px;
-}
 .swiper-item--demo {
   img {
     height: 100%;
