+++ src/swiper/demos/navigation-type.vue
@@ -1,5 +1,5 @@
 <template>
-  <t-swiper :autoplay="false" :height="180" direction="vertical" :navigation="{ type: 'fraction' }">
+  <t-swiper :autoplay="false" :navigation="{ type: 'dots-bar' }">
     <t-swiper-item class="swiper-item--demo">
       <img src="https://tdesign.gtimg.com/site/swiper/01.png" class="img" />
     </t-swiper-item>
@@ -14,15 +14,6 @@
 <script lang="ts" setup></script>
 <style lang="less">
 .swiper-item--demo {
-  width: 100%;
-  height: 180px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: #f33;
-  font-size: 20px;
-  color: #fff;
-
   img {
     height: 100%;
   }
