+++ src/swiper/swiper.vue
@@ -20,7 +20,12 @@
           <chevron-right-icon size="20px" />
         </span>
       </span>
-      <span v-if="'type' in navigation" :class="`${name}__pagination ${name}__pagination-${navigation.type || ''}`">
+      <span
+        v-if="'type' in navigation"
+        :class="`${name}__pagination ${name}__pagination-${
+          navigation.type || ''
+        } ${name}__pagination-${paginationPosition}`"
+      >
         <template v-if="['dots', 'dots-bar'].includes(navigation.type || '')">
           <span
             v-for="(item, index) in state.children.length"
@@ -104,7 +109,7 @@ export default defineComponent({
       return activeIndex + 1;
     });
     const childCount = computed(() => state.children.length);
-
+    console.log('==paginationPosition==', props.paginationPosition);
     const getContainer = (): HTMLDivElement => self?.proxy?.$el.querySelector('.t-swiper__container');
     // const getContainer = (): HTMLDivElement => swiperContainer.value as any;
     const initSwiper = () => {
